Moved components into folder, created footer w button
