+++ App.js
@@ -1,9 +1,9 @@
-import ChooseDialect from "./src/ChooseDialect";
-import Lesson from "./src/Lesson";
-import React from "react";
-import styled from "styled-components/native";
+import ChooseDialect from "./src/components/ChooseDialect";
+import Lesson from "./src/components/Lesson";
 import { NavigationContainer } from "@react-navigation/native";
+import React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
+import styled from "styled-components/native";
 
 const Stack = createStackNavigator();
 const BACKGROUND_COLOR = "#7fc8f8";
@@ -19,8 +19,8 @@ export default function App() {
             },
           }}
         >
-          <Stack.Screen name="ChooseDialect" component={ChooseDialect} />
           <Stack.Screen name="Lesson" component={Lesson} />
+          <Stack.Screen name="ChooseDialect" component={ChooseDialect} />
         </Stack.Navigator>
       </SafeareaContainer>
     </NavigationContainer>
